complete 1.2
